Moved Other refactorings
