+++ lib/Container/RpcExtension.php
@@ -18,18 +18,18 @@ use Phpactor\Extension\CodeTransform\Rpc\TransformHandler;
 use Phpactor\Extension\CodeTransform\Rpc\ClassNewHandler;
 use Phpactor\Extension\CodeTransform\Rpc\ClassInflectHandler;
 use Phpactor\Rpc\Handler\ContextMenuHandler;
-use Phpactor\Rpc\Handler\ExtractConstantHandler;
-use Phpactor\Rpc\Handler\ExtractMethodHandler;
-use Phpactor\Rpc\Handler\GenerateMethodHandler;
-use Phpactor\Rpc\Handler\GenerateAccessorHandler;
-use Phpactor\Rpc\Handler\RenameVariableHandler;
+use Phpactor\Extension\CodeTransform\Rpc\ExtractConstantHandler;
+use Phpactor\Extension\CodeTransform\Rpc\ExtractMethodHandler;
+use Phpactor\Extension\CodeTransform\Rpc\GenerateMethodHandler;
+use Phpactor\Extension\CodeTransform\Rpc\GenerateAccessorHandler;
+use Phpactor\Extension\CodeTransform\Rpc\RenameVariableHandler;
 use Phpactor\Rpc\RequestHandler\ExceptionCatchingHandler;
 use Phpactor\Rpc\RequestHandler\LoggingHandler;
 use Phpactor\Rpc\Handler\NavigateHandler;
-use Phpactor\Rpc\Handler\OverrideMethodHandler;
+use Phpactor\Extension\CodeTransform\Rpc\OverrideMethodHandler;
 use Phpactor\Rpc\Handler\CacheClearHandler;
 use Phpactor\Rpc\Handler\ConfigHandler;
-use Phpactor\Rpc\Handler\ImportClassHandler;
+use Phpactor\Extension\CodeTransform\Rpc\ImportClassHandler;
 use Phpactor\Extension\Extension;
 use Phpactor\Extension\ContainerBuilder;
 use Phpactor\Extension\Container;
