+++ lib/Extension/CodeTransform/Rpc/GenerateMethodHandler.php
@@ -1,27 +1,27 @@
 <?php
 
-namespace Phpactor\Rpc\Handler;
+namespace Phpactor\Extension\CodeTransform\Rpc;
 
-use Phpactor\CodeTransform\Domain\Refactor\GenerateAccessor;
+use Phpactor\CodeTransform\Domain\Refactor\GenerateMethod;
 use Phpactor\Rpc\Response\ReplaceFileSourceResponse;
 use Phpactor\CodeTransform\Domain\SourceCode;
+use Phpactor\Rpc\Handler\AbstractHandler;
 
-class GenerateAccessorHandler extends AbstractHandler
+class GenerateMethodHandler extends AbstractHandler
 {
-    const NAME = 'generate_accessor';
+    const NAME = 'generate_method';
     const PARAM_OFFSET = 'offset';
     const PARAM_SOURCE = 'source';
     const PARAM_PATH = 'path';
 
     /**
-     * @var GenerateAccessor
+     * @var GenerateMethod
      */
-    private $generateAccessor;
+    private $generateMethod;
 
-    public function __construct(
-        GenerateAccessor $generateAccessor
-    ) {
-        $this->generateAccessor = $generateAccessor;
+    public function __construct(GenerateMethod $generateMethod)
+    {
+        $this->generateMethod = $generateMethod;
     }
 
     public function name(): string
@@ -40,7 +40,7 @@ class GenerateAccessorHandler extends AbstractHandler
 
     public function handle(array $arguments)
     {
-        $sourceCode = $this->generateAccessor->generateAccessor(
+        $sourceCode = $this->generateMethod->generateMethod(
             SourceCode::fromStringAndPath(
                 $arguments[self::PARAM_SOURCE],
                 $arguments[self::PARAM_PATH]
