+++ lib/Extension/CodeTransform/Rpc/RenameVariableHandler.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace Phpactor\Rpc\Handler;
+namespace Phpactor\Extension\CodeTransform\Rpc;
 
 use Phpactor\CodeTransform\Domain\Refactor\RenameVariable;
 use Phpactor\Rpc\Response\ReplaceFileSourceResponse;
 use Phpactor\Rpc\Response\Input\TextInput;
 use Phpactor\Rpc\Response\Input\ChoiceInput;
 use Phpactor\CodeTransform\Domain\SourceCode;
+use Phpactor\Rpc\Handler\AbstractHandler;
 
 class RenameVariableHandler extends AbstractHandler
 {
