+++ tests/Unit/Rpc/Handler/GenerateAccessorHandlerTest.php
@@ -6,7 +6,7 @@ use Phpactor\Rpc\Handler;
 use Phpactor\CodeTransform\Domain\SourceCode;
 use Phpactor\Rpc\Response\ReplaceFileSourceResponse;
 use Phpactor\CodeTransform\Domain\Refactor\GenerateAccessor;
-use Phpactor\Rpc\Handler\GenerateAccessorHandler;
+use Phpactor\Extension\CodeTransform\Rpc\GenerateAccessorHandler;
 
 class GenerateAccessorHandlerTest extends HandlerTestCase
 {
